refactor(agentic): rename Mastra app boundary
