+++ .prettierignore
@@ -8,6 +8,6 @@ packages/graphql/src/graphql-env.d.ts
 apps/cms/schema.graphql
 apps/cms/.strapi
 apps/cms/.strapi-updater.json
-apps/mastra/.mastra
+apps/agentic/.mastra
 output/
 .cursor
