+++ apps/agentic/.env.ci
@@ -1,0 +1,10 @@
+CI=true
+NODE_ENV=test
+AGENTIC_HOST=127.0.0.1
+AGENTIC_PORT=4111
+AGENTIC_SERVICE_API_KEY=ci-placeholder
+AGENTIC_OPERATOR_API_KEY=ci-placeholder
+AGENTIC_STORAGE_URL=file:./.mastra/ci.db
+AGENTIC_MODEL=openai/gpt-5-mini
+MANAGER_BASE_URL=http://localhost:3002
+MANAGER_AGENTIC_API_KEY=ci-placeholder
