+++ apps/agentic/.env.example
@@ -1,0 +1,9 @@
+AGENTIC_HOST=127.0.0.1
+# Railway may set PORT automatically. AGENTIC_PORT takes precedence when set.
+AGENTIC_PORT=4111
+AGENTIC_SERVICE_API_KEY=replace-with-manager-to-agentic-token
+AGENTIC_OPERATOR_API_KEY=replace-with-operator-studio-token
+AGENTIC_STORAGE_URL=file:./.mastra/local.db
+AGENTIC_MODEL=openai/gpt-5-mini
+MANAGER_BASE_URL=http://localhost:3002
+MANAGER_AGENTIC_API_KEY=replace-with-agentic-to-manager-token
