+++ apps/agentic/.gitignore
