+++ apps/agentic/railway.toml
@@ -1,0 +1,10 @@
+[build]
+builder = "NIXPACKS"
+buildCommand = "pnpm --filter @forge/agentic build"
+
+[deploy]
+startCommand = "cd apps/agentic && node .mastra/output/index.mjs"
+healthcheckPath = "/health"
+healthcheckTimeout = 300
+restartPolicyType = "ON_FAILURE"
+restartPolicyMaxRetries = 3
