+++ apps/agentic/tsconfig.json
