+++ apps/agentic/vitest.config.ts
