+++ apps/manager/.env.ci
@@ -6,6 +6,6 @@ STRAPI_URL=http://localhost:1337
 STRAPI_API_TOKEN=ci-placeholder
 STRAPI_INTERNAL_API_TOKEN=ci-placeholder
 MANAGER_API_KEY=ci-placeholder
-MASTRA_BASE_URL=http://localhost:4111
-MASTRA_SERVICE_API_KEY=ci-placeholder
-MANAGER_MASTRA_API_KEY=ci-placeholder
+AGENTIC_BASE_URL=http://localhost:4111
+AGENTIC_SERVICE_API_KEY=ci-placeholder
+MANAGER_AGENTIC_API_KEY=ci-placeholder
