+++ apps/manager/.env.example
@@ -27,11 +27,11 @@ WORKFLOW_API_KEY=
 MANAGER_API_KEY=
 
 # Mastra agentic runtime — optional unless Manager calls Mastra or
-# Mastra calls Manager's /api/automations/{id}/mastra-dry-run endpoint.
-MASTRA_BASE_URL=http://localhost:4111
-MASTRA_SERVICE_API_KEY=
-MANAGER_MASTRA_API_KEY=
-MASTRA_REQUEST_TIMEOUT_MS=15000
+# Agentic calls Manager's /api/automations/{id}/agentic-dry-run endpoint.
+AGENTIC_BASE_URL=http://localhost:4111
+AGENTIC_SERVICE_API_KEY=
+MANAGER_AGENTIC_API_KEY=
+AGENTIC_REQUEST_TIMEOUT_MS=15000
 
 # Public
 NEXT_PUBLIC_WATCH_URL=
