+++ apps/manager/vitest.setup.ts
@@ -6,4 +6,4 @@ process.env.STRAPI_URL ??= "http://localhost:1337"
 process.env.STRAPI_API_TOKEN ??= "test-strapi-api-token"
 process.env.STRAPI_INTERNAL_API_TOKEN ??= "test-strapi-internal-api-token"
 process.env.MANAGER_API_KEY ??= "test-manager-api-key"
-process.env.MANAGER_MASTRA_API_KEY ??= "test-manager-mastra-api-key"
+process.env.MANAGER_AGENTIC_API_KEY ??= "test-manager-agentic-api-key"
