--- apps/mastra/.env.ci
@@ -1,10 +1,0 @@
-CI=true
-NODE_ENV=test
-MASTRA_HOST=127.0.0.1
-MASTRA_PORT=4111
-MASTRA_SERVICE_API_KEY=ci-placeholder
-MASTRA_OPERATOR_API_KEY=ci-placeholder
-MASTRA_STORAGE_URL=file:./.mastra/ci.db
-MASTRA_MODEL=openai/gpt-5-mini
-MANAGER_BASE_URL=http://localhost:3002
-MANAGER_MASTRA_API_KEY=ci-placeholder
