--- apps/mastra/.env.example
@@ -1,9 +1,0 @@
-MASTRA_HOST=127.0.0.1
-# Railway may set PORT automatically. MASTRA_PORT takes precedence when set.
-MASTRA_PORT=4111
-MASTRA_SERVICE_API_KEY=replace-with-manager-to-mastra-token
-MASTRA_OPERATOR_API_KEY=replace-with-operator-studio-token
-MASTRA_STORAGE_URL=file:./.mastra/local.db
-MASTRA_MODEL=openai/gpt-5-mini
-MANAGER_BASE_URL=http://localhost:3002
-MANAGER_MASTRA_API_KEY=replace-with-mastra-to-manager-token
