--- apps/mastra/railway.toml
@@ -1,10 +1,0 @@
-[build]
-builder = "NIXPACKS"
-buildCommand = "pnpm --filter @forge/mastra build"
-
-[deploy]
-startCommand = "cd apps/mastra && node .mastra/output/index.mjs"
-healthcheckPath = "/health"
-healthcheckTimeout = 300
-restartPolicyType = "ON_FAILURE"
-restartPolicyMaxRetries = 3
